local storage w/ pubsub start
starting local storage engine w/ pubsub support.

this is a wrapper around the default local store, but
will also have redis pubsub support for updating the
value of a feature toggle without restarting your app.

the use of pubsub will also allow you to always load from
toggle flag cache value, and only update that value
when a update is pushed. this is to reduce overhead
from querying a data store to insure flag value is
up to date.... more to come